+++ belt/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from django import get_version
 
-VERSION = (1, 0, 0, 'alpha', 1)
+VERSION = (1, 0, 0, 'final', 0)
 
 __version__ = get_version(VERSION)