+++ HISTORY.rst
@@ -3,7 +3,13 @@
 History
 -------
 
-1.0.0 (2017-2-20)
+1.0.0 (2018-10-2)
++++++++++++++++++
+
+* Updated UploadToDir with ``prefix`` param.
+* Contrib package with Django Rest Framework utils.
+
+1.0a1 (2017-2-20)
 +++++++++++++++++
 
 * First release on PyPI.